Switch duration type

+++ fluo/arduino_program/arduino_program.ino
@@ -88,7 +88,7 @@ void shuffle (uint8_t* array, int arrayLength) {
         array[randomIndex] = temporaryValue;
     }
 }
-void bars(uint16_t _duration) {
+void bars(uint32_t _duration) {
     unsigned long duration = _duration * 1000;
     unsigned long t = millis();
 
@@ -113,7 +113,7 @@ void bars(uint16_t _duration) {
     }
 }
 
-void snake(uint16_t _duration) {
+void snake(uint32_t _duration) {
     unsigned long duration = _duration * 1000;
     unsigned long t = millis();
 
@@ -187,7 +187,7 @@ void snake(uint16_t _duration) {
     }
 }
 
-void randomFade(uint16_t _duration, uint16_t topHold) {
+void randomFade(uint32_t _duration, uint16_t topHold) {
     unsigned long duration = _duration * 1000;
     unsigned long t = millis();
 
@@ -216,7 +216,7 @@ void randomFade(uint16_t _duration, uint16_t topHold) {
     }
 }
 
-void drops(uint16_t _duration) {
+void drops(uint32_t _duration) {
     unsigned long duration = _duration * 1000;
     unsigned long t = millis();
 
@@ -347,7 +347,7 @@ void plotLine(int8_t _x0, int8_t _y0, int8_t _x1, int8_t _y1, int8_t state) {
     }
 }
 
-void spokes(uint16_t _duration, uint8_t invert) {
+void spokes(uint32_t _duration, uint8_t invert) {
     unsigned long duration = _duration * 1000;
     unsigned long t = millis();
